feat: begin moving onMint over to ScrollCampaign
